elements_projects/flaskProject/requirements.txt
# pip install -r requirements.txt
flask
python-bitcoinrpc
flask-cors
simplejson
